--- simulator/covid19
@@ -1,1 +1,0 @@
-../covid19
