+++ simulator/covid19/utils.py
